+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.70.1] — 2026-08-28
+
 ### Fixed
 
 - **A view's `Parameters` section no longer loses the parent's `@param` prose